feat: 메인 페이지 section 분리
