+++ zoo-client/src/components/home/index.ts
@@ -1,0 +1,7 @@
+'use client';
+
+export { default as FirstSection } from './FirstSection';
+export { default as SecondSection } from './SecondSection';
+export { default as ThirdSection } from './ThirdSection';
+export { default as FourthSection } from './FourthSection';
+export { default as FifthSection } from './FifthSection';
